refactor: Remove unused fields and update data types in schemas
This commit removes the `expires_in` field from the `Token` schema in order to simplify the data structure. Additionally, it updates the data types of the `user_id` and `rev` fields in the `Setting` schema to use UUID and int respectively, improving data consistency. The `DetailSchema` is also added to the `schemas` module.